third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/declared/declared.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Declared StylePropertyMap does not contain inline styles
  assert_equals: expected (undefined) undefined but got (object) null
[FAIL] Declared StylePropertyMap does not contain properties with invalid values
  assert_equals: expected (undefined) undefined but got (object) null
Harness: the test ran to completion.
